travel_agent/pyproject.toml
[tool.poetry]
name = "travel-example"
version = "0.1.0"
description = "A travel agent example built with Rosetta."
repository = "https://github.com/couchbaselabs/rosetta-example"
authors = [
    "Glenn Galvizo <glenn.galvizo@couchbase.com>"
]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"

# For downloading and ingesting blog data into Couchbase.
newspaper3k = "^0.2.8"
semchunk = "^2.1.0"
lxml-html-clean = "^0.1.1"
sentence-transformers = "^3.0.1"
python-dotenv = "^1.0.1"
couchbase = "^4.3.0"

# For hosting servers (in general).
fastapi = "^0.111.1"

# For building a sample agent interface.
requests = "^2.32.3"
streamlit = "^1.36.0"
streamlit-feedback = "^0.1.3"

# The Rosetta project (the front-facing parts)!
agent-catalog = { git = "git@github.com:couchbaselabs/rosetta.git", extras = ["langchain"] }
# agent-catalog = { path = PATH_TO_LOCAL_ROSETTA, extras = ["langchain"], develop = true }

[tool.poetry.group.dev.dependencies]
pre-commit = "^3.0.0"

[tool.poetry.group.controlflow]
optional = true

[tool.poetry.group.controlflow.dependencies]
controlflow = { git = "git@github.com:PrefectHQ/ControlFlow.git" }

[tool.ruff]
line-length = 120
lint.select = ["F", "B", "I", "SIM"]
show-fixes = true

[tool.ruff.lint.isort]
no-sections = true
force-single-line = true
lines-between-types = 1
force-wrap-aliases = true
split-on-trailing-comma = false

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
